+++ packages/bpk-component-input/index.js
@@ -17,8 +17,9 @@
  */
 
 import BpkInput, { INPUT_TYPES as TYPES } from './src/BpkInput';
+import CLEAR_BUTTON_MODES from './src/clearButtonModes';
 import withOpenEvents from './src/withOpenEvents';
 
 export default BpkInput;
 export const INPUT_TYPES = TYPES;
-export { withOpenEvents };
+export { withOpenEvents, CLEAR_BUTTON_MODES };